♻️ いいね機能に関するリファクタリング
